refactor(cognitarium): move query variables elements
