+++ contracts/okp4-cognitarium/src/querier/mod.rs
@@ -1,3 +1,8 @@
 mod engine;
 mod plan;
 mod plan_builder;
+mod variable;
+
+pub use engine::*;
+pub use plan::*;
+pub use plan_builder::*;
